refactor(factories): consolidate factory files from 14 to 7
Merged factory files to reduce duplication and improve navigation:
- pipeline_factory.py: GenericPipelineFactory + runner assembly
- services_factory.py: BaseServicesFactory + ServicesBuilder + RunnerServices
- adapters_factory.py: HttpClientFactory + DataSourceFactory + DataSourceRegistry
- registry.py: pipeline factory instances + transformer registry

Deleted:
- clients.py (empty)
- base_services_factory.py, services_builder.py, runner_services.py (merged)
- http_client_factory.py, data_sources.py, data_source_registry.py (merged)
- generic_factory.py, runner_assembly.py, pipeline_factories.py (merged)
- transformer_factory.py (merged)

Updated imports across codebase (source, tests, docs).
