+++ docs/03-guides/registry-pattern.md
@@ -54,7 +54,7 @@ if ProviderRegistry.is_registered("chembl"):
 Thin facade over ProviderRegistry for backward compatibility. Located in `src/bioetl/composition/factories/data_source_registry.py`.
 
 ```python
-from bioetl.composition.factories.data_source_registry import DataSourceRegistry
+from bioetl.composition.factories.adapters_factory import DataSourceRegistry
 
 # Old way (still works, delegates to ProviderRegistry)
 creator = DataSourceRegistry.get("chembl")
