+++ tests/infrastructure/factories/test_data_sources.py
@@ -6,7 +6,7 @@ from unittest.mock import Mock
 
 import pytest
 
-from bioetl.composition.factories.data_sources import DataSourceFactory
+from bioetl.composition.factories.adapters_factory import DataSourceFactory
 from bioetl.infrastructure.adapters.http.client import UnifiedHTTPClient
